#169 Updated: Unified vendor-main/mini-headers
Updated the vendor-main-header.php and vendor-mini-header.php to include
some new useful variables, as well as unify the variables so that each
template has the same available to each other.  Exception: mini has a
reference to the shop URL, not needed in main since main is already the
same page as the URL would be.